requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=1.4
pymysql>=1.0
pydantic>=1.10.12
email-validator>=1.3
alembic>=1.8
python-dotenv>=1.0

# Testing & dev
pytest>=7.2
pytest-cov>=4.0
httpx>=0.24
pre-commit>=3.4
black>=24.3
ruff>=0.26
bandit>=1.7
